requirements.txt
torch
nuscenes-devkit
numpy
matplotlib
pypdf
scikit-image
scipy
shapely
